requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
PyYAML
tqdm

# Terminal.app

## local package installation
## this will treate and istaall "src" folder as package in local system
## uses local setup.py file
-e .